examples: improve examples with a small README and additional clarity to follow along with what they are supposed to be doing.
