+++ examples/chat/README.md
@@ -1,0 +1,15 @@
+# chat
+
+Uses `yzma` to create a simple chat program.
+
+## Running
+
+```shell
+go run ./examples/chat/ -model ./models/qwen2.5-0.5b-instruct-fp16.gguf
+```
+
+## Install
+
+```shell
+go install ./examples/chat
+```
