+++ examples/describe/README.md
@@ -2,13 +2,15 @@
 
 Describes an image using a Vision Language Model. Works with either remote or local files.
 
-```shell
-describe https://www.publicdomainpictures.net/pictures/220000/t2/llama-1496506141lWt.jpg 2>/dev/null
-```
-
 ## Install
 
 ```
-go install .
+go install ./examples/describe
+```
+
+## Running
+
+```shell
+describe https://www.publicdomainpictures.net/pictures/220000/t2/llama-1496506141lWt.jpg 2>/dev/null
 ```
