+++ examples/hello/README.md
@@ -1,0 +1,9 @@
+# hello
+
+The "hello, world" of `yzma`.
+
+## Running
+
+```shell
+go run ./examples/hello/ 2>/dev/null
+```
